x86/microcode/intel: Drop wbinvd() for Intel CPUs when loading microcode
Some older processors had a bad interaction when updating microcode if the
caches were dirty causing machine checks. The wbinvd() was added to
mitigate that before performing microcode updates. Now that Linux checks
for the minimum version before performing an update, those microcode
revisions can't be loaded.

Early loading is also not required to use wbinvd() any longer, that was
added as a safety net.

Remove calls to wbinvd().

Signed-off-by: Ashok Raj <ashok.raj@intel.com>