x86/microcode: Add a facility to control microcode update flows
Apart from the microcode_ops structure members, add a simple facility to
control behavior, or requirements in some cases. For AMD microcode its
required to run the update on both threads of the core, while this isn't
required for the Intel microcode loader.

Provide a facility to declare support for control. Which will be used by
the core loader subsystem.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>