x86/microcode: Skip NMI rendezvous when Uniform update is supported
When processor reports supporting Uniform microcode update, there is no
reason to do perform the NMI late dance. Uniform update provides
architectural gaurantees that the update will not be triggered in the
middle of an instruction execution.

Set a control flag to indicate to indicate that NMI late dance is not
required.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>
Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>